pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "An e-commerce backend service"
authors = [
    {name = "Precious Udoibok"}
]
readme = "README.md"
requires-python = "3.13.2"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.15.0"
autoflake = "^2.3.1"
isort = "^6.0.1"
alembic = "^1.15.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
